pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "cloud-services-backup-cli"
version = "0.1"
description = "A command-line interface for backing up data from popular cloud services to a local server"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.9"
dependencies = [
    "python-dateutil==2.9.*",
    "pillow==11.3.*",
    "pillow_heif==1.1.*",
    "piexif==1.1.*",
    "hachoir==3.3.*",
]

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
cloud-service-backup = "cloud_services_backup_cli.__main__:main"
